+++ app/classes/data.py
@@ -68,22 +68,6 @@ class Post(Document):
         'ordering': ['-createdate']
     }
 
-# Refercnce Teacher Class
-class Comment(Document):
-    # Line 63 is a way to access all the information in Course and Teacher w/o storing it in this class
-    author = ReferenceField('User',reverse_delete_rule=CASCADE) 
-    post = ReferenceField('Post',reverse_delete_rule=CASCADE)
-    # This could be used to allow comments on comments
-    # comment = ReferenceField('Comment',reverse_delete_rule=CASCADE)
-    # Line 68 is where you store all the info you need but won't find in the Course and Teacher Object
-    content = StringField()
-    create_date = DateTimeField(default=dt.datetime.utcnow)
-    modify_date = DateTimeField()
-    role = StringField("Role")
-
-    meta = {
-        'ordering': ['-createdate']
-    }
 
 # _____________________ iRate
 
@@ -109,6 +93,24 @@ class TeacherCourse(Document):
     course_files = FileField()
     create_date = DateTimeField(default=dt.datetime.utcnow)
     modify_date = DateTimeField()
+
+    meta = {
+        'ordering': ['-createdate']
+    }
+
+    # Refercnce Teacher Class
+class Comment(Document):
+    # Line 63 is a way to access all the information in Course and Teacher w/o storing it in this class
+    author = ReferenceField('User',reverse_delete_rule=CASCADE) 
+    # post = ReferenceField('Post',reverse_delete_rule=CASCADE)
+    course = ReferenceField('Courses',reverse_delete_rule=CASCADE)
+    # This could be used to allow comments on comments
+    # comment = ReferenceField('Comment',reverse_delete_rule=CASCADE)
+    # Line 68 is where you store all the info you need but won't find in the Course and Teacher Object
+    content = StringField()
+    create_date = DateTimeField(default=dt.datetime.utcnow)
+    modify_date = DateTimeField()
+    role = StringField("Role")
 
     meta = {
         'ordering': ['-createdate']
