+++ app/templates/commentform.html
@@ -10,8 +10,8 @@
             {% endfor %}
         {% endfor %}
 
-        <h1 class="display-5">{{post.subject}}</h1>
-        {{post.content}} <br>
+        <h1 class="display-5">{{course.course_title}}</h1>
+        {{course.course_number}} <br>
         <h1 class="display-5">New Comment</h1>
 
         <form method=post>
