+++ app/templates/index.html
@@ -10,7 +10,9 @@
         <div class="ml-5 py-3" style="width: 50%; margin-left: 25%">
                 <div class="input-group px-5">
                     <input type="search" class="form-control rounded" placeholder="Search" aria-label="Search" aria-describedby="search-addon" />
+                    <a href="/course/list">
                     <button type="button" class="btn btn-primary">search</button>
+                    </a>
                 </div>
         </div>   
         <p class="font-italic py-3 text-gradient">Search by
